Implementing i2c port 1 write function for OLED driver

+++ firmware/src/system/include/peripheral/ssd1306/ssd1306.h
@@ -66,7 +66,7 @@ typedef struct {
 //  Function definitions
 //
 
-uint8_t ssd1306_Init();
+int ssd1306_Init();
 void ssd1306_UpdateScreen();
 void ssd1306_Fill(SSD1306_COLOR color);
 void ssd1306_DrawPixel(uint8_t x, uint8_t y, SSD1306_COLOR color);
